NodeResourceTopology: Make `default` as the default namespace in config
In order to enable Topology aware scheduling in a Kuberetes cluster, we need to
deploy a producer which exposes resource topology information as CRs and the
NodeResourceTopology scheduler plugin.

An upstream producer of the NodeResourceTopology instances is Node Feature Discovery.
The associated PR where this feature is enabled in NFD is here:
https://github.com/kubernetes-sigs/node-feature-discovery/pull/525

The default namespace where NFD deploys CR instances is `default` as can be seen here:
 https://github.com/kubernetes-sigs/node-feature-discovery/blob/master/cmd/nfd-master/main.go#L108
so in order to have a better user experience and to make sure that both
these components work together seamlessly we change the default namespace
in the scheduler plugin config to be the same as in NFD i.e. `default`.

Signed-off-by: Swati Sehgal <swsehgal@redhat.com>