NodeResourceTopology: handle best-effort pods in score plugin
Scheduling best-effort pods with score plugin leads to scheduler pod crashing.
Best-effort pods are not requesting any numeric value of resources so all resources are 0, thus program panics with a division by zero operation.

For every Best-effort pod we should return a score of 100 since from the NodeResourceTopology perspective, every node is a good fit for best-effort pods and this is what this patch is aiming to do.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>