noderesourcetopology: add more tests and internal cleanups (#356)
* nrt: filter: add multi-container tests

Increase the coverage of the noderesourcetopology filter plugin
adding more multi-container tests.
We want to test with non-uniform containers (e.g. containers
with different resource requirements).
We start with the simple scope=pod and we will expand later.

Signed-off-by: Francesco Romani <fromani@redhat.com>

* nrt: filter: flip the return value of resMatchNUMANodes

As emerged in
https://github.com/kubernetes-sigs/scheduler-plugins/pull/347#discussion_r814564195
the code becomes a bit more readable if we make resMatchNUMANodes
return true if the requested resources do fit in any available
NUMA zone.

Signed-off-by: Francesco Romani <fromani@redhat.com>

* nrt: filter: rename function for clarity

Now that resMatchNUMANodes return true if the requested resources
fit in any provided NUMA zones, we can rename to a bit more expressive
name to gain a tad more of readability.

Trivial rename only: no intended changes in behaviour.

Signed-off-by: Francesco Romani <fromani@redhat.com>

* nrt: doc: document the NRT fundamental prerequisite

A fundamental assumption for NRT to work correctly is that
all the worker nodes must have a consistent NUMA topology
and kubelet configuration.

While the former prerequisite is not so strict and can be relaxed,
a heteregenous host HW can lead to surprising (but not wrong) scheduling
decisions, so we don't recommend it at this point in time.

On the other hand, a consistent kubelet configuration it's a
prerequisite, because it can lead to a bad scheduling decision.

Unfortunately, it's very hard to detect inconsistencies at scheduling
run time, and it's also unfeasible to enforce these prerequisite -
the scheduler (or the NodeResourceTopology CRD!) is just the wrong
place to do that.

So, the only option left is to explicitely document this
prerequisite and point cluster administrators to existing kubernetes
tools which will enable them to deal with mixed HW topology and/or
mixed kubelet configuration.

Signed-off-by: Francesco Romani <fromani@redhat.com>