noderesourcetopology: Addtional tests
Since in the previous commit we perform a check to evaluate that a
resource is a device if it is a non-native resource, we want to make
sure that we are not causing unexpected regressions when it comes to
how we handle other extended resources. Extended resources can be
resources known to the node but not available to be allocated on a
NUMA node in cases where devices don't have a topology or where they
are custom resources attached to specific nodes for custom use
cases.

This patch adds additional test cases to make sure we have more
comprehensive tests to tackle different combinations of resource
request when it comes to pods that request devices and extended
resources.

Signed-off-by: Swati Sehgal <swsehgal@redhat.com>