nrt: cache: reserve: detect foreign pods
When designing the caching code for the noderesourcetopology (nrt)
plugin, a key decision was to assume that the set of scheduling domains,
was clearly defined and never overlapping.
We call "scheduling domain" the set of nodes a given scheduler (default,
NUMA-aware) can schedule workload to.

This restriction greatly simplifies the design, but turned out too
fragile. Noteworthy examples (not exhaustive list):
1. the default/primary scheduler still manages, and must manage, the
   non-workload pods. We must tolerate pod restarts, node restart
   and so on.
2. CI and test environment can require the default/primary scheduler
   to send pods to create the testing condition

So, the NUMA-aware/secondary scheduler needs to be more robust and to
detect the case on which the primary scheduler sent some pods, and resync
the resource availability.

Lacking better names so far, we call this state as "foreign pods".
Meaning a node has "foreign pods" if pods are detected running on that
node, but they were scheduled by a different scheduler

Since we watch all pods anyway -we need to-, the core idea is to
monitor pods not scheduled by ourselves (the pod.Spec.SchedulerName is
handy here) landing on nodes we are supposed to manage.
We detect the nodes we are supposed to manager checking if we have
NodeTopologyInfo about them; if so, we own them; if not, we ignore them.
If we detect unexpected traffic to these nodes, we mark them "tainted",
considering them unsuitable (no resources) for scheduling until a resync
iteration happens.

Signed-off-by: Francesco Romani <fromani@redhat.com>