nrt: reserve: track reserved but unbounded pods
The current code tracking pods locally expects near-zero time
between the reserve stage and the binding stage.
If this holds true, the code is meant to handle gracefully events
which trigger a resync, because this means that is guaranteed
that NRT updates are either consistent or can be safely ignored.

Let's consider this scenario:
1. pod1, pod2 already scheduled and running. We are at steady state:
   NRT info from node matches pfpExpected (= PFP computed locally by the scheduler)
2. pod3 comes, passes reserve stage, resources are overreserved.
   pod3 experiences unexpected delays in bonding stage
3. pod4 comes. The scheduler correctly overreserved, and pod4 is
   unschedulable, because the node is filtered out.
   However, pod3 is not running, so is not considered in the expected pod set.
4. a NRT update is pushed because regular timer fired. pod3 is not yet
   running, so its PFP considers only pod1 and pod2
5. a resync happens, while pod3 is not yet bound.
   The resync loop checks expected pods on both sides, and they match.
   The flush happens. The observed NRT is committed, and this causes no
   harm. But the overreservation are flushed, and that's wrong!
6. pod4 is retried, and because overreservation have been flushed, it is
   scheduled. TopologyAffinityError will happen.

A fix for this scenario is to start tracking pod right after the
succesful reserve stage, not after the bonding stage (= when we receive
a update with non-empty nodeName).

While this change per se is easy, care must be taken to handle deletions,
in this scenario, and not to leave stale data in the caches.

Signed-off-by: Francesco Romani <fromani@redhat.com>