nrt: generalize topology manager config options
The NodeResourceTopology API is deprecating the `TopologyPolicies`
field, which was already a debatable way to express the Topology Manager
configuration.
Future version of the API will add top-level attributes, which can
encode the required TopologyManager configuration in key-value format.

An alternative legal way (which some implementations may also prefer)
is to provide an additional special-purpose Zone whose Type is
"TopologyManager". The scheduler will ignore all the data but the
Attributes.

Signed-off-by: Francesco Romani <fromani@redhat.com>