nrt: overreserve: use pfp compute method from NRT
When computing the expected node pod fingerprint (PFP), we
now look for the attribute describing the computation method
provided by the topology-updater agent. If not provided, we
fall back to the previous backward-compatible method, which
is consider all pods.

If the method is supplied, the scheduler plugin now computes
the PFP in the same way advertised as the node agent.

Finally, in order to keep the user fully in control, add
config tunables to force the global behavior of the resync
loop. The cluster admin can let the plugin autodetect
(default behavior) or force one of the existing methods.

Signed-off-by: Francesco Romani <fromani@redhat.com>