nrt: cache: add support for dedicated informer
Add an option to create and use a separate informer
to get unfiltered pod listing from the apiserver.

Due to mismatched view with respect to the kubelet,
the plugin needs access to pod in terminal phase
which are not deleted to make sure the reconciliation
is done correctly.

xref: https://github.com/kubernetes-sigs/scheduler-plugins/issues/598
xref: https://github.com/kubernetes/kubernetes/issues/119423

Signed-off-by: Francesco Romani <fromani@redhat.com>