nrt: score: add tests for partial NRT data
Currently, the NRT (NodeResourceTopology) filter plugin
just exits and allows silently nodes which don't have NRT
objects available. This was done back in time to allow
the plugin to coexist with the regular scheduler, and in general
work in mixed scenarios on which only some node have NRT data
reported. IOW, the decision wants to enable clusters with
non-homogeneous nodes.

The score plugins need to work in these scenarios as well.
It needs to tolerate gracefully partial NRT data availability
(some nodes have, some nodes have not) and should always prefer
nodes which have NRT data over nodes which don't.

In this patch we start tests for these scenarios. More to come.

Signed-off-by: Francesco Romani <fromani@redhat.com>