docs: anyone should be any
Problem: the comment and associated docs should say if
there are not resources available needed for the group,
the scheduler should not start any.
Solution: update all references anyone->any.

It could also be stated "not start any one" but
"not start any" is a more natural way to say it

Signed-off-by: vsoch <vsoch@users.noreply.github.com>