nrt: rewrite resource accounting for scope=container
Rewrite the accounting of NUMA-local resources when
using scope=container. The previous code was too lenient
and worked mostly by side effects when dealing with
non-NUMA affine resources.

A non-NUMA affine resource (aka a hostlevel resource)
is a resource which is not guaranteed to always have
a NUMA affinity. CPU and memory (incl. hugepages) always do,
but devices may or may not, both options are legal for
device plugins.

Similarly, ephemeral storage is a prominent example of resource
which should never have a NUMA affinity.
The accounting in this case was wrong because previously the
resource was considered NUMA affine.

Note: it's legal to configure topology updaters (e.g. NFD)
to not advertise CPU and memory in NRT objects.
Thus is best to treat lack of them as warnings, not
as blocking errors.

However if the per-NUMA affine counters go negative
this is definitely an error condition we need to detect
and be very loud about it.

Signed-off-by: Francesco Romani <fromani@redhat.com>