nrt: resync NRT data also on attribute change
One of the key assumptions we took when designing the
NodeResourceTopology (NRT) plugin is that the kubelet
configuration of the worker node changes *very* rarely,
if at all, during the cluster lifetime.
As rule of thumb, it was expected to change with a
frequency of like once every quarter (3 months) or so,
and likely less often. So the event of changing during
a scheduling cycle was deemed extremely low.

However, we fail to notice kubelet configuration changes
(the bits we care reported by NRT data) and we update
them incidentally when resyncing the cache.

These updates are expected to be rare, but still failing
to noticing them is a  much worse issue: with out of
date configuration information, the scheduler plugin
will take wrong decisions until restarted.

Up until now, the mitigation was to restart the scheduler
once kubelet config changes; this works, but it is
unpractical and requires more orchestration.

We add the option to resync the NRT data if the attribute
change (and nothing else did) to overcome this limitation.
In the current framework, because of how controller-runtime/client
works, this will require a new connection to the apiserver to
watch the changes and react to them, adding the node
to the set of the one being resynced in the next resync iteration.

Even considering HA scheduler scenarios, this will cause
a very limited amount of new connections, and should not
cause any scalability concern. Nevertheless, we make
the feature opt-in.

Signed-off-by: Francesco Romani <fromani@redhat.com>