nrt: enable consumption of MaxAllowedNUMANodes
kubernetes 1.31 gained support for topology manager options,
and most notably to bump the limit of max NUMA nodes,
which is no longer hardcoded to 8.

Thus, we need to adapt the scheduler code.
Likewise other key topology manager parameters,
we trust NRT objects to carry the values we need.
In this case, to mitigate DoS risks, we set an artificial
"high enough" (and hopefully good enough for a while)
limit.

Once obtained the vlaue, we pass it through to all
the code places on which the previous hardcoded value was used.

Signed-off-by: Francesco Romani <fromani@redhat.com>